Copy TodoItemTaskInput to AddTodoItem
Replace TodoItemTaskInput with (new) AddTodoItem in TodoList
